django-oauth/requirements.txt
django
pika
django-cors-headers
PyJWT